requirements.txt
python-telegram-bot==13.15
telethon
transformers
datasets
torch
pandas
scikit-learn
seqeval
shap
lime
python-dotenv
matplotlib
notebook